refactor Node, and Device ConnectionFailListeners to their own class files, and updated relevant javadocs
