+++ app/src/main/java/com/idevicesinc/sweetblue/P_BondManager.java
@@ -5,7 +5,6 @@ import static com.idevicesinc.sweetblue.BleDeviceState.BONDING;
 import static com.idevicesinc.sweetblue.BleDeviceState.UNBONDED;
 
 import com.idevicesinc.sweetblue.BondListener.BondEvent;
-import com.idevicesinc.sweetblue.BleDevice.ConnectionFailListener;
 import com.idevicesinc.sweetblue.BondListener.Status;
 import com.idevicesinc.sweetblue.BleDeviceConfig.BondFilter;
 import com.idevicesinc.sweetblue.UhOhListener.UhOh;
@@ -128,11 +127,11 @@ final class P_BondManager
 		return false;
 	}
 	
-	Object[] getOverrideBondStatesForDisconnect(ConnectionFailListener.Status connectionFailReasonIfConnecting)
+	Object[] getOverrideBondStatesForDisconnect(DeviceConnectionFailListener.Status connectionFailReasonIfConnecting)
 	{
 		final Object[] overrideBondingStates;
 		
-		if( connectionFailReasonIfConnecting == ConnectionFailListener.Status.BONDING_FAILED )
+		if( connectionFailReasonIfConnecting == DeviceConnectionFailListener.Status.BONDING_FAILED )
 		{
 			overrideBondingStates = OVERRIDE_UNBONDED_STATES;
 		}
@@ -162,7 +161,7 @@ final class P_BondManager
 		{
 			final boolean doingReconnect_shortTerm = m_device.is(BleDeviceState.RECONNECTING_SHORT_TERM);
 			
-			m_device.disconnectWithReason(BleDevice.ConnectionFailListener.Status.BONDING_FAILED, status.timing(), BleStatuses.GATT_STATUS_NOT_APPLICABLE, failReason, m_device.NULL_READWRITE_EVENT());
+			m_device.disconnectWithReason(DeviceConnectionFailListener.Status.BONDING_FAILED, status.timing(), BleStatuses.GATT_STATUS_NOT_APPLICABLE, failReason, m_device.NULL_READWRITE_EVENT());
 		}
 		else
 		{
