+++ app/src/main/java/com/idevicesinc/sweetblue/P_ServerConnectionFailEntry.java
@@ -65,7 +65,7 @@ final class P_ServerConnectionFailEntry
 
 		final int ePlease__PE_Please = m_mngr.invokeCallback(e);
 
-		if( BleNode.ConnectionFailListener.Please.isRetry(ePlease__PE_Please) )
+		if( NodeConnectionFailListener.Please.isRetry(ePlease__PE_Please) )
 		{
 			m_mngr.m_server.connect_internal(nativeDevice);
 		}
