+++ app/src/main/java/com/idevicesinc/sweetblue/P_ServerConnectionFailManager.java
@@ -70,7 +70,7 @@ final class P_ServerConnectionFailManager
 		{
 			final BleServer.ConnectionFailListener.Please please = m_connectionFailListener.onEvent(e);
 
-			ePlease__PE_Please = please != null ? please.please() : BleNode.ConnectionFailListener.Please.PE_Please_DO_NOT_RETRY;
+			ePlease__PE_Please = please != null ? please.please() : NodeConnectionFailListener.Please.PE_Please_DO_NOT_RETRY;
 		}
 		else
 		{
